Large update to site UI
Adds a colour scheme, logo, header, other css niceities for the move
viewer portion as well.

Minor tweaks to how the move viewer handles, including making sure that
the play controls are not shown when they are not useable.

Cleaned up the code a bit and removed unused and unnecessary things.