Bugfix for development server
Turns off hot reloading for the /public/ directory because it was being
dumb and refreshing on AJAX calls to the move data files.

As part of the investigation the AJAX request was moved up and out of
MoveInfo and its been left there as it makes for a nicer workflow.