New theme fixes + go-live + bugfixes
Moved around some of the .css so that useful styles like headers are
propagated through the whole site via index.css.

Added an unfinished About.js / Footer, which are disabled for now.

Moved the template into the src folder so it won't get published and be
accessible.

Fixed bug with editing the frame textbox directly where it would skip
frames.