Lots more changes
Adds "About" page content.

Changes "Footer" to "Menu" and places it at the top of the page instead
of the bottom.

Adds a button to copy a link to the move, mainly to inform the user in a
nice way that they can link any page of the app to anyone to show them
what they're seeing.

Now calculates invincible and superarmor frame ranges, and handles
local intangible frames (i.e. ones that don't come from a param file).
If any of these have no range, instad of showing N/A they are just not
shown at all to save space.